Move /login_guest and /login_previous into system_bridge agent_commands
Removes the commands from actual_web.rs and registers them in
SystemBridgePlugin's AgentCommandsPlugin, consistent with other
system-level commands.
