+++ crates/system_bridge/src/agent_commands.rs
@@ -8,9 +8,50 @@ pub struct AgentCommandsPlugin;
 
 impl Plugin for AgentCommandsPlugin {
     fn build(&self, app: &mut App) {
+        app.add_console_command::<LoginGuestCommand, _>(login_guest_cmd);
+        app.add_console_command::<LoginPreviousCommand, _>(login_previous_cmd);
         app.add_console_command::<LogoutCommand, _>(logout_cmd);
         app.add_console_command::<LiveScenesCommand, _>(live_scenes_cmd);
         app.add_console_command::<ChatCommand, _>(chat_cmd);
+    }
+}
+
+// --- /login_guest ---
+
+/// Login as guest (session-only, profile will not persist)
+#[derive(clap::Parser, ConsoleCommand)]
+#[command(name = "/login_guest")]
+struct LoginGuestCommand;
+
+fn login_guest_cmd(
+    mut input: ConsoleCommand<LoginGuestCommand>,
+    mut events: EventWriter<SystemApi>,
+) {
+    if let Some(Ok(_)) = input.take() {
+        events.write(SystemApi::LoginGuest);
+        input.reply_ok("logging in as guest");
+    }
+}
+
+// --- /login_previous ---
+
+/// Login using previously saved credentials
+#[derive(clap::Parser, ConsoleCommand)]
+#[command(name = "/login_previous")]
+struct LoginPreviousCommand;
+
+fn login_previous_cmd(
+    mut input: ConsoleCommand<LoginPreviousCommand>,
+    mut events: EventWriter<SystemApi>,
+    mut pending: ResMut<PendingConsoleResponses>,
+) {
+    if let Some(Ok(_)) = input.take() {
+        let (sx, rx) = common::rpc::RpcResultSender::<Result<(), String>>::channel();
+        events.write(SystemApi::LoginPrevious(sx));
+        pending.push_receiver(rx, |result| match result {
+            Ok(()) => Ok("logged in with previous credentials".to_string()),
+            Err(e) => Err(e),
+        });
     }
 }
